backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.12
alembic>=1.10.4
psycopg2-binary>=2.9.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Telegram API
telethon>=1.28.5
cryptography>=42.0.0

# Machine Learning
scikit-learn>=1.2.2
numpy>=1.24.3
joblib>=1.2.0
scipy>=1.10.1

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.2  # For database connection retries
aiohttp>=3.8.5  # For async HTTP requests 